requirements.txt
# Install the Kosmos 2.5 PR branch of Transformers.
git+https://github.com/tic-top/transformers.git@main

torch
transformers
pillow
gradio
accelerate
torchvision